third_party/blink/web_tests/fast/css/content-language-only-whitespace-expected.txt

Test for bug 76701: map HTTP-EQUIV content-language to -webkit-locale. This particular test tests that a content-language with whitespace-only content is ignored. This expectation may change, see bug. HTML5 decrees that the meta element be ignored in case of whitespace only content. It's unclear what other browsers do.

FAIL languageOfNode('x') should be auto. Was "  \9 \a  ".
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
